requirements.txt
# Core dependencies
psycopg2-binary==2.9.10
python-dotenv==1.0.1
PyYAML==6.0.2

# Web interface
streamlit==1.39.0
streamlit-option-menu==0.3.12

# Data processing
pandas==2.2.3
numpy==1.26.4
beautifulsoup4==4.12.3
python-frontmatter==1.1.0

# Database
sqlalchemy==2.0.36
alembic==1.13.3

# HTTP requests
requests==2.32.4
httpx==0.27.2

# Async support
asyncio==3.4.3
aiofiles==24.1.0

# Logging and utilities
structlog==24.4.0
click==8.1.7
tqdm==4.67.1

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1

# Optional: For Claude imports
anthropic==0.40.0

# Optional: For image processing
Pillow==11.1.0

# Optional: For embeddings
openai==1.58.1
